requirements.txt
pytest
requests
pandas
psycopg2-binary
sqlalchemy
python-dotenv
schedule